+++ yogaposeguide/src/main/java/com/example/yogaposeguide/repository/RegisterDetailsRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RegisterDetailsRepository extends JpaRepository<RegisterDetails,Integer> {
+public interface RegisterDetailsRepository extends JpaRepository<RegisterDetails,Long> {
     Optional<RegisterDetails> findByUserName(String username);
 }
