+++ yogaposeguide/src/main/java/com/example/yogaposeguide/repository/YogaPoseRepository.java
@@ -3,5 +3,5 @@ package com.example.yogaposeguide.repository;
 import com.example.yogaposeguide.models.YogaPoses;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface YogaPoseRepository extends JpaRepository<YogaPoses,Integer> {
+public interface YogaPoseRepository extends JpaRepository<YogaPoses,Long> {
 }
